Introducing GridContainer hierarchy
