+++ src/main/java/edu/ie3/datamodel/models/input/container/GridContainer.java
@@ -1,0 +1,124 @@
+package edu.ie3.datamodel.models.input.container;
+
+import edu.ie3.datamodel.exceptions.AggregationException;
+import edu.ie3.datamodel.models.UniqueEntity;
+import edu.ie3.datamodel.models.input.NodeInput;
+import edu.ie3.datamodel.models.voltagelevels.VoltageLevel;
+
+import java.util.*;
+import java.util.stream.Collectors;
+
+public class GridContainer implements InputContainer {
+    /** Name of this grid */
+    protected final String gridName;
+    /** Accumulated raw grid elements (lines, nodes, transformers, switches) */
+    protected final RawGridElements rawGrid;
+    /** Accumulated system participant elements */
+    protected final SystemParticipantElements systemParticipants;
+    /** Accumulated graphic data entities (node graphics, line graphics) */
+    protected final GraphicElements graphics;
+
+    public GridContainer(String gridName, RawGridElements rawGrid, SystemParticipantElements systemParticipants, GraphicElements graphics) {
+        this.gridName = gridName;
+        this.rawGrid = rawGrid;
+        this.systemParticipants = systemParticipants;
+        this.graphics = graphics;
+    }
+
+    @Override
+    public void add(UniqueEntity entity) {
+        try {
+            rawGrid.add(entity);
+            return;
+        } catch (IllegalArgumentException ignored) { // No further exception handling needed
+        }
+        try {
+            systemParticipants.add(entity);
+            return;
+        } catch (IllegalArgumentException ignored) { // No further exception handling needed
+        }
+        try {
+            graphics.add(entity);
+            return;
+        } catch (IllegalArgumentException ignored) { // No further exception handling needed
+        }
+        throw new IllegalArgumentException(
+                "Entity type is unknown, cannot add entity [" + entity + "]");
+    }
+
+    @Override
+    public List<UniqueEntity> allEntitiesAsList() {
+        List<UniqueEntity> allEntities = new LinkedList<>();
+        allEntities.addAll(rawGrid.allEntitiesAsList());
+        allEntities.addAll(systemParticipants.allEntitiesAsList());
+        allEntities.addAll(graphics.allEntitiesAsList());
+        return Collections.unmodifiableList(allEntities);
+    }
+
+    @Override
+    public boolean areValuesValid() {
+        if (!rawGrid.areValuesValid()) return false;
+        if (!systemParticipants.areValuesValid()) return false;
+        return graphics.areValuesValid();
+    }
+
+    public String getGridName() {
+        return gridName;
+    }
+
+    public RawGridElements getRawGrid() {
+        return rawGrid;
+    }
+
+    public SystemParticipantElements getSystemParticipants() {
+        return systemParticipants;
+    }
+
+    public GraphicElements getGraphics() {
+        return graphics;
+    }
+
+    /**
+     * Determining the predominant voltage level in this grid by counting the occurrences of the
+     * different voltage levels
+     *
+     * @param rawGrid Raw grid elements
+     * @return The predominant voltage level in this grid
+     * @throws AggregationException If not a single, predominant voltage level can be determined
+     */
+    protected static VoltageLevel determinePredominantVoltLvl(RawGridElements rawGrid)
+            throws AggregationException {
+        return rawGrid.getNodes().stream()
+                .map(NodeInput::getVoltLvl)
+                .collect(Collectors.groupingBy(voltLvl -> voltLvl, Collectors.counting()))
+                .entrySet()
+                .stream()
+                .max(Map.Entry.comparingByValue())
+                .map(Map.Entry::getKey)
+                .orElseThrow(
+                        () -> new AggregationException("Cannot determine the predominant voltage level."));
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        GridContainer that = (GridContainer) o;
+        return gridName.equals(that.gridName) &&
+                rawGrid.equals(that.rawGrid) &&
+                systemParticipants.equals(that.systemParticipants) &&
+                graphics.equals(that.graphics);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(gridName, rawGrid, systemParticipants, graphics);
+    }
+
+    @Override
+    public String toString() {
+        return "GridContainer{" +
+                "gridName='" + gridName + '\'' +
+                '}';
+    }
+}
