+++ src/main/java/edu/ie3/datamodel/models/input/container/SubGridContainer.java
@@ -6,27 +6,14 @@
 package edu.ie3.datamodel.models.input.container;
 
 import edu.ie3.datamodel.exceptions.AggregationException;
-import edu.ie3.datamodel.models.UniqueEntity;
-import edu.ie3.datamodel.models.input.NodeInput;
 import edu.ie3.datamodel.models.voltagelevels.VoltageLevel;
-import java.util.*;
-import java.util.stream.Collectors;
 
 /** Represents the accumulation of all data needed to create a complete single grid */
-public class SubGridContainer implements InputContainer {
-
-  /** Name of this grid */
-  private final String gridName;
+public class SubGridContainer extends GridContainer {
   /** subnet number of this grid */
   private final int subnet;
   /** Predominantly apparent voltage level in this single grid */
   private final VoltageLevel predominantVoltageLevel;
-  /** Accumulated raw grid elements (lines, nodes, transformers, switches) */
-  private final RawGridElements rawGrid;
-  /** Accumulated system participant elements */
-  private final SystemParticipantElements systemParticipants;
-  /** Accumulated graphic data entities (node graphics, line graphics) */
-  private final GraphicElements graphics;
 
   public SubGridContainer(
       String gridName,
@@ -35,11 +22,8 @@ public class SubGridContainer implements InputContainer {
       SystemParticipantElements systemParticipants,
       GraphicElements graphics)
       throws AggregationException {
-    this.gridName = gridName;
+    super(gridName, rawGrid, systemParticipants, graphics);
     this.subnet = subnet;
-    this.rawGrid = rawGrid;
-    this.systemParticipants = systemParticipants;
-    this.graphics = graphics;
 
     try {
       this.predominantVoltageLevel = determinePredominantVoltLvl(rawGrid);
@@ -54,46 +38,6 @@ public class SubGridContainer implements InputContainer {
     }
   }
 
-  @Override
-  public void add(UniqueEntity entity) {
-    try {
-      rawGrid.add(entity);
-      return;
-    } catch (IllegalArgumentException ignored) { // No further exception handling needed
-    }
-    try {
-      systemParticipants.add(entity);
-      return;
-    } catch (IllegalArgumentException ignored) { // No further exception handling needed
-    }
-    try {
-      graphics.add(entity);
-      return;
-    } catch (IllegalArgumentException ignored) { // No further exception handling needed
-    }
-    throw new IllegalArgumentException(
-        "Entity type is unknown, cannot add entity [" + entity + "]");
-  }
-
-  @Override
-  public List<UniqueEntity> allEntitiesAsList() {
-    List<UniqueEntity> allEntities = new LinkedList<>();
-    allEntities.addAll(rawGrid.allEntitiesAsList());
-    allEntities.addAll(systemParticipants.allEntitiesAsList());
-    allEntities.addAll(graphics.allEntitiesAsList());
-    return Collections.unmodifiableList(allEntities);
-  }
-
-  @Override
-  public boolean areValuesValid() {
-    if (!rawGrid.areValuesValid()) return false;
-    if (!systemParticipants.areValuesValid()) return false;
-    return graphics.areValuesValid();
-  }
-
-  public String getGridName() {
-    return gridName;
-  }
 
   public int getSubnet() {
     return subnet;
@@ -103,36 +47,12 @@ public class SubGridContainer implements InputContainer {
     return predominantVoltageLevel;
   }
 
-  public RawGridElements getRawGrid() {
-    return rawGrid;
-  }
-
-  public SystemParticipantElements getSystemParticipants() {
-    return systemParticipants;
-  }
-
-  public GraphicElements getGraphics() {
-    return graphics;
-  }
-
-  /**
-   * Determining the predominant voltage level in this grid by counting the occurrences of the
-   * different voltage levels
-   *
-   * @param rawGrid Raw grid elements
-   * @return The predominant voltage level in this grid
-   * @throws AggregationException If not a single, predominant voltage level can be determined
-   */
-  private static VoltageLevel determinePredominantVoltLvl(RawGridElements rawGrid)
-      throws AggregationException {
-    return rawGrid.getNodes().stream()
-        .map(NodeInput::getVoltLvl)
-        .collect(Collectors.groupingBy(voltLvl -> voltLvl, Collectors.counting()))
-        .entrySet()
-        .stream()
-        .max(Map.Entry.comparingByValue())
-        .map(Map.Entry::getKey)
-        .orElseThrow(
-            () -> new AggregationException("Cannot determine the predominant voltage level."));
+  @Override
+  public String toString() {
+    return "SubGridContainer{" +
+            "gridName='" + gridName + '\'' +
+            ", subnet=" + subnet +
+            ", predominantVoltageLevel=" + predominantVoltageLevel +
+            '}';
   }
 }
